Move sed keys related logic in another partial plugin
